Increase the number and verbosity of compiler warnings
- Enable several additional compiler warnings
- Treat warnings as errors
- Fix a number of issues detected through the new, more aggressive warnings
- Add weakify/strongify macros from libextobjc for easier memory management of self in blocks
- Remove several deprecated method implementations
- Add a new date helpers object to vend a specific style of NSDateFormatter that we use regularly

+++ Libraries/aaronbrethorst-onebusaway-iphone-common-fa9a21b/Classes/View/OBAProgressIndicatorView.h
@@ -24,7 +24,7 @@
 
 - (id) initWithFrame:(CGRect)frame;
 
-- (void) setMessage:(NSString*)message inProgress:(BOOL)inProgress progress:(float)progress;
-- (void) setInProgress:(BOOL)inProgress progress:(float)progress;
+- (void) setMessage:(NSString*)message inProgress:(BOOL)inProgress progress:(CGFloat)progress;
+- (void) setInProgress:(BOOL)inProgress progress:(CGFloat)progress;
 
 @end
